requirements.txt
pillow
pytesseract
streamlit